Analyzing Crime in Los Angeles
Los Angeles, California, attracts people from all over the world who are looking to be successful and make a name for themselves. And with that comes a lot of opportunities, not always of the good kind! While LA has a reputation for beautiful weather and a laid back lifestyle, it's also been renowned for a high volume of crime, which is not particularly surprising given it is the second most populous US city.

In this project you'll serve as a data detective, supporting the Los Angeles Police Department (LAPD) in analyzing their crime data to help shape how they should allocate resources to best protect the people of their city!